requirements.txt
Django
django-bootstrap4
pyopenssl
djangorestframework